cr_extraction/requirements.txt
requests==2.31.0
MechanicalSoup==1.3.0
beautifulsoup4==4.12.2
Flask==2.3.2
functions-framework==3.4.0
Pillow==10.0.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
img2pdf==0.4.4